clean up components, add lazy loading
