+++ anees_uzzaman_choudhury/coding_task/frontend/package.json
@@ -23,6 +23,7 @@
     "react-hot-toast": "^2.4.1",
     "react-icons": "^5.2.1",
     "react-redux": "^9.1.2",
+    "react-spinners": "^0.13.8",
     "redux-thunk": "^3.1.0"
   },
   "devDependencies": {
